init koa
