+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "my-web-task-assist-react",
+  "version": "1.0.0",
+  "description": "my web task assist react",
+  "main": "index.js",
+  "directories": {
+    "lib": "lib"
+  },
+  "scripts": {
+    "start": "SET NODE_ENV=development&& nodemon index.js",
+    "build": "node build.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/rejiejay/myweb-task-assist.git"
+  },
+  "keywords": [
+    "react"
+  ],
+  "author": "rejiejay",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/rejiejay/myweb-task-assist/issues"
+  },
+  "homepage": "https://github.com/rejiejay/myweb-task-assist#readme",
+  "dependencies": {
+    "koa": "^2.11.0"
+  }
+}
